+++ src/nu/mem/tree/__init__.py
@@ -1,8 +1,0 @@
-"""nu-mem tree meta-passes: ref inlining."""
-
-from .inline_refs import inline_refs
-
-
-__all__ = [
-    "inline_refs",
-]